requirements.txt
frappe
azure-storage